+++ app/controllers/events_controller.rb
@@ -1,6 +1,5 @@
 class EventsController < ApplicationController
-  before_action :logged_in_user, only: [:checkin, :create]
-  before_action :find_event, only: [:checkin, :assistents]
+  before_action :logged_in_user, only: [:create]
 
   def index
     @eventspast = Event.past
@@ -24,29 +23,10 @@ class EventsController < ApplicationController
     end
   end
 
-  def checkin
-    attendance = @event.assistent_relationships.build(user_id: current_user.id, role: 'assistent')
-
-    if attendance.save
-      flash[:success] = 'Check-in Successfully'
-      redirect_to :root
-    else
-      flash[:danger] = 'Error in Check-in'
-    end
-  end
-
-  def assistents
-    @assistents = @event.assistents
-  end
-
   private
 
   def event_params
     params.require(:event).permit(:title, :schedule, :description)
-  end
-
-  def find_event
-    @event = Event.find(params[:event_id])
   end
 
 end
