+++ app/helpers/attendances_helper.rb
@@ -1,0 +1,2 @@
+module AttendancesHelper
+end
