+++ app/views/attendances/show.html.erb
