+++ config/routes.rb
@@ -1,14 +1,15 @@
 Rails.application.routes.draw do
   root 'events#index'
 
-  get       '/signup',  to: 'users#new'
-  post      '/signup',  to: 'users#create'
-  get       '/login',   to: 'sessions#new'
-  post      '/login',   to: 'sessions#create'
-  delete    '/logout',  to: 'sessions#destroy'
-  resources 'users',    only: [:new, :create, :show]
-  resources 'sessions', only: [:new]
-  resources 'events',   only: [:index, :new, :create] do
+  get       '/signup',   to: 'users#new'
+  post      '/signup',   to: 'users#create'
+  get       '/login',    to: 'sessions#new'
+  post      '/login',    to: 'sessions#create'
+  delete    '/logout',   to: 'sessions#destroy'
+  resources 'users',     only: [:new, :create, :show]
+  resources 'sessions',  only: [:new]
+  resources 'events',    only: [:index, :new, :create]
+  resources 'attendance' do
     post :checkin
     get  :assistents
   end
