+++ spec/controllers/attendances_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe AttendancesController, type: :controller do
+
+end
